جابجایی کامند های یوزر برای جدا سازی اونا از کامند های فرندشیپ
