+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/AddUserAddress/AddUserAddressCommandRespone.cs
@@ -1,4 +1,4 @@
-﻿namespace MakFood.Customer.Application.Commands.AddUserAddress
+﻿namespace MakFood.Customer.Application.Commands.User.AddUserAddress
 {
     public class AddUserAddressCommandRespone
     {
