+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/AddUserAddress/AddUserAddressMapper.cs
@@ -1,5 +1,5 @@
 ﻿using MakFood.Customer.Domain.UserAggregate;
-namespace MakFood.Customer.Application.Commands.AddUserAddress
+namespace MakFood.Customer.Application.Commands.User.AddUserAddress
 {
     public static class AddUserAddressMapper
     {
