+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/RegisterUser/RegisterUserCommandResponse.cs
@@ -1,4 +1,4 @@
-﻿namespace MakFood.Customer.Application.Commands.RegisterUser
+﻿namespace MakFood.Customer.Application.Commands.User.RegisterUser
 {
     public record RegisterUserCommandResponse
     {
