+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/RemoveUserAddress/RemoveUserAddressCommandHandler.cs
@@ -3,7 +3,7 @@ using MakFood.Customer.Domain.UserAggregate.Contracts;
 using MakFood.Customer.Infrastructure.Persistence.Context.Transactions;
 using MediatR;
 
-namespace MakFood.Customer.Application.Commands.RemoveUserAddress
+namespace MakFood.Customer.Application.Commands.User.RemoveUserAddress
 {
     public class RemoveUserAddressCommandHandler : IRequestHandler<RemoveUserAddressCommand, RemoveUserAddressCommandRespone>
     {
