+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/RemoveUserAddress/RemoveUserAddressCommandRespone.cs
@@ -1,4 +1,4 @@
-﻿namespace MakFood.Customer.Application.Commands.RemoveUserAddress
+﻿namespace MakFood.Customer.Application.Commands.User.RemoveUserAddress
 {
     public class RemoveUserAddressCommandRespone
     {
