+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/UpdateUserAddress/UpdateUserAddressCommandRespone.cs
@@ -1,4 +1,4 @@
-﻿namespace MakFood.Customer.Application.Commands.UpdateUserAddress
+﻿namespace MakFood.Customer.Application.Commands.User.UpdateUserAddress
 {
     public class UpdateUserAddressCommandRespone
     {
