+++ MakFood.Customer/MakFood.Customer.Application/Commands/User/UpdateUserInformation/UpdateUserInformationCommandValidator.cs
@@ -1,6 +1,6 @@
 ﻿using FluentValidation;
 
-namespace MakFood.Customer.Application.Commands.UpdateUserInformation
+namespace MakFood.Customer.Application.Commands.User.UpdateUserInformation
 {
     public class UpdateUserInformationCommandValidator : AbstractValidator<UpdateUserInformationCommand>
     {
